refactor : story 파일 위치 각 컴포넌트 폴더 내부로 변경
